moved queue and stack to a utility file
